Address review feedback for static split fix
Agent-Logs-Url: https://github.com/Chetansahney/pytorch-forecasting/sessions/bd3ddb3a-a457-48aa-aad4-de2eafd763bb

+++ pytorch_forecasting/data/_tslib_data_module.py
@@ -221,7 +221,7 @@ class _TslibDataset(Dataset):
         if processed_data["static"] is not None:
             static_features = processed_data["static"]
             if not isinstance(static_features, torch.Tensor):
-                static_features = torch.tensor(static_features)
+                static_features = torch.tensor(static_features, dtype=torch.float32)
             static_features = static_features.flatten()
 
             static_feature_names = metadata["feature_names"]["static"]
